requirements.txt
streamlit
pillow
transformers
torch
langchain
langchain-core
langchain-google-genai
python-dotenv
requests
